# Vulkan sync rewrite