This PR reworks the synchronization in the Vulkan backend to allow pushing render commands ahead of the GPU, similarly to the existing backends.

Currently there's a CPU stall after every compute shader, texture/buffer transfer and render pass (all draw calls of one view). This results in high CPU frame time not only because of the stall (CPU does nothing while GPU is busy) but also because of the non-negligible cost of submitting command buffers. For some examples this easily wastes > 1ms of CPU time.

The frame latency can be controlled with [`Init::Resolution::maxFrameLatency`](https://bkaradzic.github.io/bgfx/bgfx.html#_CPPv4N4bgfx10Resolution15maxFrameLatencyE), from 1 to 3. Using 0 there defaults to 3, similar to the D3D11/12 backends. Current behaviour is similar to a fixed latency of 1, only ever allowing 1 frame to be encoded or executed at the same time. A latency of 2 allows one frame to be encoded on the CPU while another frame is rendering on the GPU, etc.

I would appreciate if anyone could test this with systems I have no access to. The only environment I could test on was an Nvidia card on Windows 10. While the validation layers are happy, there might be subtle bugs in the swapchain interaction that just didn't surface for me.

Sorry for the huge changelist, but this touched just about every part of the code. Looking at single commits might make the changes easier to follow.

### Numbers

Some measurements with example 09-hdr (Release, Nvidia GTX 1070, Win 10):

#### Old code:

| Function | # of calls | combined CPU time |
| --- |---:| ---:|
| `vkQueueSubmit` | 17 | 0.4ms |
| `vkQueueWaitIdle` | 17 | 1.2ms |
| `vkWaitForFences` | 0 | -- |
| Total | -- | 1.6ms |

#### New code, frame latency 3 (default):

Note: there are two submissions because the example uses texture readback.

| Function | # of calls | combined CPU time |
| --- |---:| ---:|
| `vkQueueSubmit` | 2 | 0.1ms |
| `vkQueueWaitIdle` | 0 | -- |
| `vkWaitForFences` | 4 | 0.03ms |
| Total | -- | 0.2ms |

#### New code, frame latency 1:

| Function | # of calls | acc. CPU time |
| --- |---:| ---:|
| `vkQueueSubmit` | 2 | 0.1ms |
| `vkQueueWaitIdle` | 0 | -- |
| `vkWaitForFences` | 4 | 0.3ms |
| Total | -- | 0.4ms |

### Extra fixes

Because I was already working on it:

- Swapchain images are not acquired if no view in the frame renders to the backbuffer. Potentially fixes https://github.com/bkaradzic/bgfx/issues/1936
- Textures can be assigned a debug name now
- Cubemap blitting from and to faces > 0 was fixed

### Related Todo

Possible future fixes:

- Rework swapchain recreation to retire all swapchain-related items and release them deferred, instead of stalling the GPU. This should make resizing a lot smoother.

### Details

Some more info for anyone curious:

Frames are encoded in a ring buffer of command buffers (in `CommandQueueVK`). An entire frame is recorded to a single command buffer, so there's only ever one `vkQueueSubmit` per frame (with the exception of texture/screenshot readback). Frames are serialized by making each submission wait on the previous submission's semaphore. This has the useful side effect of introducing a full memory barrier between consecutive frames.

Before rendering the frame a single memory barrier makes sure that all transfer operations (texture and buffer creation/updates) are finished and visible. Render passes are synchronized to other render passes and compute dispatches via external subpass dependencies (roughly `ALL_COMMANDS` -> `GRAPHICS` -> `ALL_COMMANDS`). Since subpass dependencies don't cover compute dispatches, we need to serialize those with another memory barrier between them.

The barriers and dependencies could, in a future change, be made more fine-grained, depending on things like framebuffer attachment flags, etc.